docker/docker_compose_basic/web/requirements.txt
Flask==0.12
mysql-connector==2.1.6